+++ ja2lib/Strategic/StrategicMap.h
@@ -25,7 +25,7 @@ extern INT8 gbWorldSectorZ;
 
 #define NUMBER_OF_SAMS 4
 
-extern INT16 pSamList[NUMBER_OF_SAMS];
+extern u8 pSamList[NUMBER_OF_SAMS];
 extern INT16 pSamGridNoAList[NUMBER_OF_SAMS];
 extern INT16 pSamGridNoBList[NUMBER_OF_SAMS];
 
@@ -52,8 +52,8 @@ extern BOOLEAN gfUseAlternateMap;
 #define CHECK_DIR_X_DELTA (WORLD_TILE_X * 4)
 #define CHECK_DIR_Y_DELTA (WORLD_TILE_Y * 10)
 
-void GetCurrentWorldSector(INT16 *psMapX, INT16 *psMapY);
-BOOLEAN SetCurrentWorldSector(INT16 sMapX, INT16 sMapY, INT8 bMapZ);
+void GetCurrentWorldSector(u8 *psMapX, u8 *psMapY);
+BOOLEAN SetCurrentWorldSector(u8 sMapX, u8 sMapY, i8 bMapZ);
 
 BOOLEAN EnterSector(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
 void UpdateMercsInSector(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
@@ -62,7 +62,7 @@ void UpdateMercInSector(struct SOLDIERTYPE *pSoldier, u8 sSectorX, u8 sSectorY, 
 void InitializeMapStructure();
 
 // get short sector name without town name
-void GetShortSectorString(INT16 sMapX, INT16 sMapY, STR16 sString, size_t bufSize);
+void GetShortSectorString(u8 sMapX, u8 sMapY, STR16 sString, size_t bufSize);
 
 // NEW!
 // Calculates the name of the sector based on the loaded sector values.
@@ -72,7 +72,7 @@ void GetShortSectorString(INT16 sMapX, INT16 sMapY, STR16 sString, size_t bufSiz
 // equivalent)
 void GetLoadedSectorString(STR16 pString, size_t bufSize);
 
-void GetMapFileName(INT16 sMapX, INT16 sMapY, INT8 bSectorZ, STR8 bString, BOOLEAN fUsePlaceholder,
+void GetMapFileName(u8 sMapX, u8 sMapY, INT8 bSectorZ, STR8 bString, BOOLEAN fUsePlaceholder,
                     BOOLEAN fAddAlternateMapLetter);
 
 // Called from within tactical.....
@@ -117,7 +117,7 @@ void PrepareLoadedSector();
 // handle for slay...no better place to really put this stuff
 void HandleSlayDailyEvent(void);
 
-void HandleQuestCodeOnSectorEntry(INT16 sNewSectorX, INT16 sNewSectorY, INT8 bNewSectorZ);
+void HandleQuestCodeOnSectorEntry(u8 sNewSectorX, u8 sNewSectorY, INT8 bNewSectorZ);
 
 // handle a soldier leaving thier squad behind, this sets them up for mvt and potential rejoining of
 // group
@@ -134,7 +134,7 @@ BOOLEAN IsThereAFunctionalSAMSiteInSector(u8 sSectorX, u8 sSectorY, INT8 bSector
 BOOLEAN IsSectorDesert(u8 sSectorX, u8 sSectorY);
 
 // sam site under players control?
-INT32 SAMSitesUnderPlayerControl(INT16 sX, INT16 sY);
+INT32 SAMSitesUnderPlayerControl(u8 sX, u8 sY);
 
 INT8 GetSAMIdFromSector(u8 sSectorX, u8 sSectorY, INT8 bSectorZ);
